# plans/openapi-worker.yaml
openapi: 3.0.3
info:
  title: SpecTrace Worker API
  version: 1.0.0
  description: |-
    The v1 contract the Cloudflare Worker implements. Every path under `/api/v1/` survives from the Django server with its schemas intact, minus the Linear and GitHub webhook integrations. Five endpoints are new: the CLI pushes the requirement tree (`PUT /api/v1/specs/`), impact reports (`POST /api/v1/results/impact/`), and drift reports (`POST /api/v1/results/drift/`); `spectrace lore sync` drains the task outbox (`GET /api/v1/tasks/outcomes`, `POST /api/v1/tasks/outcomes/ack`). `GET /api/v1/specs/impact/` and `GET /api/v1/specs/drift/` serve the last pushed report instead of computing one; their response shapes stay.

    The Worker enforces the API key on every route: send it in the `X-API-Key` header. A missing or wrong key returns 401 with an `ErrorEnvelope`.

    Two error shapes exist. Routes that moved from `api.py` keep the legacy `ErrorResponse` (`{"success": false, "error": "..."}`); every other route returns `ErrorEnvelope` (`{"error": {"code", "message", "details"}}`). Success bodies on those routes wrap the payload as `{"data": ..., "meta": ...}` with `meta` present only when the route paginates.
servers:
- url: /
security:
- apiKey: []
...
paths:
  /api/v1/specs/:
    get:
      tags:
      - Specs
      summary: Page one project's requirements in tree order
      description: |-
        Lists requirements ordered by materialized path, so parents precede their children. `parent_id` narrows the page to that requirement's descendants; naming a requirement the store lacks returns an empty page rather than an error. `tags` takes a comma-separated list and matches a requirement carrying any of them. `status`, `verification_status`, and `risk_level` each match one value exactly. `max_depth` limits how deep below the scope the page reaches: `1` returns the roots, or the direct children of `parent_id`. Every row carries a `descendants` rollup counting its whole subtree by verification status, with the highest risk level found there, so a collapsed parent still shows what lies beneath it. `attention=1` keeps only requirements that are not passing or that carry high or critical risk. When `parent_id` names a stored requirement, `data.scope` carries it with its ancestors, root first, for breadcrumbs.
      operationId: api-v1-specs_get
      parameters:
      - $ref: '#/components/parameters/ProjectParam'
      - name: status
        in: query
        description: Lifecycle status, matched exactly
        schema:
          type: string
      - name: verification_status
        in: query
        schema:
          $ref: '#/components/schemas/VerificationStatus'
      - name: risk_level
        in: query
        schema:
          $ref: '#/components/schemas/RiskLevel'
      - name: tags
        in: query
        description: Comma-separated tags; a requirement carrying any of them matches
        schema:
          type: string
      - name: parent_id
        in: query
        description: External id whose descendants make up the page
        schema:
          type: string
      - name: max_depth
        in: query
        description: Levels below the scope to include; 1 returns the roots, or the direct children of `parent_id`
        schema:
          type: integer
          minimum: 1
      - name: attention
        in: query
        description: '`1` keeps only requirements that are not passing or that carry high or critical risk'
        schema:
          type: string
          enum:
          - '1'
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: One page of requirements
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - meta
                properties:
                  data:
                    $ref: '#/components/schemas/RequirementPage'
                  meta:
                    $ref: '#/components/schemas/PagePosition'
        '400':
          description: Ambiguous project; `details.projects` lists the candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags:
      - Specs
      summary: Replace one project's requirement tree, test links, and flows
      description: |-
        Bulk upsert from `spectrace push`. With `replace` the payload is the whole project: requirements the store holds for `project` but the payload omits are deleted, as are their test links and the flows the payload omits; without it, omitted rows stay. Requirements key on `external_id`, links on `(test_nodeid, requirement_id)`, flows on `name`. The tree is rebuilt from `parent_id`, siblings ordered by `external_id`, so a requirement whose parent changed moves. References in `parent_id`, `depends_on`, `links[].requirement_id`, and `flows[].requirements` resolve against the payload first and then against stored requirements of other projects; a reference that resolves nowhere is dropped and listed in `unresolved`. `verification_status`, `slo_status`, and `structure_completeness` are computed by the Worker and not accepted.
      operationId: api-v1-specs_put
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecPushRequest'
      responses:
        '200':
          description: Counts per kind after the store matches the payload
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/SpecPushResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/specs/coverage/:
    get:
      tags:
      - Specs
      summary: Coverage metrics and stale requirements for one project
      description: |-
        Counts requirements by verification status. `stale_requirements` are the requirement ids named by `stale_link` errors and `spec_drift` warnings in the last drift report pushed for the project, intersected with the project's requirements; without a stored report the list is empty.
      operationId: api-v1-specs-coverage_get
      parameters:
      - $ref: '#/components/parameters/ProjectParam'
      responses:
        '200':
          description: Coverage report
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/CoverageReport'
        '400':
          description: Ambiguous project; `details.projects` lists the candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/specs/coverage/trend:
    get:
      tags:
      - Specs
      summary: The coverage snapshot series for one project
      description: |-
        Serves the snapshots `POST /api/v1/results/coverage/` stored, oldest first, so a chart reads left to right. `limit` keeps the most recent runs. The Worker computes nothing here; `spectrace specs coverage --push` measures the checkout and sends the numbers.
      operationId: api-v1-specs-coverage-trend_get
      parameters:
      - $ref: '#/components/parameters/ProjectParam'
      - name: limit
        in: query
        description: Most recent snapshots to return (default 100, max 500)
        schema:
          type: integer
          default: 100
          maximum: 500
      responses:
        '200':
          description: Coverage trend
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/CoverageTrend'
        '400':
          description: Ambiguous project; `details.projects` lists the candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/specs/drift/:
    get:
      tags:
      - Specs
      summary: The last pushed drift report
      description: |-
        Serves the report most recently stored by `POST /api/v1/results/drift/` for the project. The Worker computes nothing here; `spectrace drift --push` runs detection where the repo lives. Response shape unchanged from the Django server, plus `stored_at`.
      operationId: api-v1-specs-drift_get
      parameters:
      - $ref: '#/components/parameters/ProjectParam'
      responses:
        '200':
          description: Drift report
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/StoredDriftReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No drift report has been pushed for the project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/specs/impact/:
    get:
      tags:
      - Specs
      summary: The last pushed impact report
      description: |-
        Serves the report most recently stored by `POST /api/v1/results/impact/` for the project. The Worker computes nothing here; `spectrace impact --push` runs the analysis against git where the repo lives. The `spec_id` and `file_path` query parameters are gone with the computation. Response shape unchanged from the Django server, plus the report's `base`, `head`, `generated_at`, `stored_at`, and optional `code` block.
      operationId: api-v1-specs-impact_get
      parameters:
      - $ref: '#/components/parameters/ProjectParam'
      responses:
        '200':
          description: Impact report
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/StoredImpactReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No impact report has been pushed for the project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/specs/{external_id}/context:
    get:
      tags:
      - Specs
      summary: Surrounding context for one requirement
      description: |-
        The requirement with its test links, dependencies, dependents, and FRET fields. No trailing slash.
      operationId: api-v1-specs-context_get
      parameters:
      - $ref: '#/components/parameters/ExternalIdParam'
      responses:
        '200':
          description: Requirement context
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/SpecContext'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/specs/{external_id}/status/:
    get:
      tags:
      - Specs
      summary: Verification status for one requirement
      operationId: api-v1-specs-status_get
      parameters:
      - $ref: '#/components/parameters/ExternalIdParam'
      responses:
        '200':
          description: Requirement status
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                properties:
                  data:
                    $ref: '#/components/schemas/SpecStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: The `SPECTRACE_API_KEY` Worker secret. Required on every route.
  parameters:
    ProjectParam:
      name: project
      in: query
      description: Project name; defaults to the deployment's `SPECTRACE_PROJECT`
      schema:
        type: string
    ExternalIdParam:
      name: external_id
      in: path
      required: true
      schema:
        type: string
...
    PageParam:
      name: page
      in: query
      description: 1-based page number; a page past the last returns the last page
      schema:
        type: integer
        default: 1
        minimum: 1
    PerPageParam:
      name: per_page
      in: query
      description: Page size (max 100). A non-integer value is a 400.
      schema:
        type: integer
        default: 25
        maximum: 100
  responses:
    Unauthorized:
      description: Missing or wrong `X-API-Key`; `code` is `unauthorized`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Malformed JSON, missing field, or invalid query parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: No such resource; `code` is `not_found`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    RiskLevel:
      type: string
      enum:
      - unclassified
      - low
      - medium
      - high
      - critical
    VerificationMethod:
      type: string
      enum:
      - unspecified
      - test
      - inapp
      - both
    VerificationStatus:
      type: string
      enum:
      - untested
      - passing
      - failing
...
    SpecPushRequest:
      type: object
      required:
      - project
      - requirements
      - links
      - flows
      properties:
        project:
          type: string
          minLength: 1
        replace:
          type: boolean
          default: false
          description: Delete the project's stored requirements, their test links, and the flows the payload omits
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/RequirementUpsert'
        links:
          type: array
          items:
            $ref: '#/components/schemas/TestLinkUpsert'
        flows:
          type: array
          items:
            $ref: '#/components/schemas/FlowUpsert'
...
    TestLinkUpsert:
      description: One entry of `.spectrace/links.json` as `extract_links` writes it.
      type: object
      required:
      - test_nodeid
      - requirement_id
      properties:
        test_nodeid:
          type: string
          maxLength: 500
        requirement_id:
          type: string
    FlowUpsert:
      description: One flow YAML document as `YAMLFlowParser` reads it.
      type: object
      required:
      - name
      - display_name
      - steps
      properties:
        name:
          type: string
          maxLength: 100
          description: The YAML `id`
        display_name:
          type: string
          maxLength: 200
          description: The YAML `title`
        description:
          type: string
          default: ''
        version:
          type: integer
          default: 1
        source_file:
          type: string
          default: ''
        requirements:
          type: array
          items:
            type: string
          default: []
        steps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FlowStepDefinition'
...
    UpsertCounts:
      type: object
      required:
      - created
      - updated
      - deleted
      properties:
        created:
          type: integer
        updated:
          type: integer
        deleted:
          type: integer
...
    SpecPushResult:
      type: object
      required:
      - project
      - requirements
      - links
      - flows
      - unresolved
      properties:
        project:
          type: string
        requirements:
          $ref: '#/components/schemas/UpsertCounts'
        links:
          $ref: '#/components/schemas/UpsertCounts'
        flows:
          $ref: '#/components/schemas/UpsertCounts'
        unresolved:
          type: array
          description: References dropped because they name no requirement in the payload or the store
          items:
            type: object
            required:
            - kind
            - source
            - ref
            properties:
              kind:
                type: string
                enum:
                - parent_id
                - depends_on
                - link
                - flow
              source:
                type: string
                description: The requirement, test nodeid, or flow name holding the reference
              ref:
                type: string
                description: The requirement id that resolved nowhere
    RequirementPage:
      type: object
      required:
      - project
      - requirements
      properties:
        project:
          type: string
        scope:
          $ref: '#/components/schemas/RequirementScope'
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/RequirementSummary'
    RequirementScope:
      type: object
      description: The `parent_id` requirement and its ancestors, root first; present only when `parent_id` names a stored requirement
      required:
      - external_id
      - title
      - ancestors
      properties:
        external_id:
          type: string
        title:
          type: string
        ancestors:
          type: array
          items:
            type: object
            required:
            - external_id
            - title
            properties:
              external_id:
                type: string
              title:
                type: string
    RequirementSummary:
      type: object
      required:
      - external_id
      - title
      - description
      - path
      - depth
      - numchild
      - tags
      - priority
      - status
      - verification_status
      - verification_method
      - risk_level
      - source_file
      - descendants
      properties:
        external_id:
          type: string
        title:
          type: string
        description:
          type: string
        path:
          type: string
          description: Materialized path; sorting by it yields tree order
        depth:
          type: integer
        numchild:
          type: integer
        tags:
          type: array
          items:
            type: string
        priority:
          type: string
        status:
          type: string
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        verification_method:
          $ref: '#/components/schemas/VerificationMethod'
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        source_file:
          type: string
        descendants:
          $ref: '#/components/schemas/DescendantRollup'
...
    PagePosition:
      type: object
      required:
      - page
      - per_page
      - total
      - total_pages
      - has_next
      - has_prev
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    CoverageReport:
      type: object
      required:
      - project
      - metrics
      - stale_requirements
      properties:
        project:
          type: string
        metrics:
          type: object
          required:
          - total
          - non_draft
          - passing
          - failing
          - untested
          - stale
          properties:
            total:
              type: integer
            non_draft:
              type: integer
            passing:
              type: integer
            failing:
              type: integer
            untested:
              type: integer
            stale:
              type: integer
        stale_requirements:
          type: array
          items:
            type: string
...
    CoverageTrend:
      type: object
      required:
      - project
      - snapshots
      properties:
        project:
          type: string
        snapshots:
          type: array
          description: Oldest first
          items:
            $ref: '#/components/schemas/CoverageSnapshot'
...
    SpecContext:
      type: object
      required:
      - external_id
      - title
      - description
      - tags
      - status
      - verification_status
      - priority
      - test_results
      - depends_on
      - depended_by
      properties:
        external_id:
          type: string
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        priority:
          type: string
        test_results:
          type: array
          items:
            type: object
            required:
            - test_nodeid
            - last_status
            properties:
              test_nodeid:
                type: string
              last_status:
                type: string
        depends_on:
          type: array
          items:
            type: string
        depended_by:
          type: array
          items:
            type: string
        fret:
          type: object
          description: Present only when at least one FRET field is set; each key present only when set
          properties:
            scope:
              type: string
            condition:
              type: string
            component:
              type: string
            timing:
              type: string
            response:
              type: string
    SpecStatus:
      type: object
      required:
      - external_id
      - title
      - verification_status
      - last_checked
      - latest_result
      - regression
      properties:
        external_id:
          type: string
        title:
          type: string
        verification_status:
          $ref: '#/components/schemas/VerificationStatus'
        last_checked:
          type: string
          format: date-time
          nullable: true
        latest_result:
          type: object
          nullable: true
          required:
          - status
          - message
          - checked_at
          - steps_passed
          - steps_failed
          properties:
            status:
              type: string
            message:
              type: string
            checked_at:
              type: string
              format: date-time
            steps_passed:
              type: integer
            steps_failed:
              type: integer
        regression:
          type: object
          required:
          - is_regression
          properties:
            is_regression:
              type: boolean
            previous_status:
              type: string
              nullable: true
            regressed_at:
              type: string
              format: date-time
              nullable: true
...
    StoredImpactReport:
      allOf:
      - $ref: '#/components/schemas/ImpactReport'
      - type: object
        required:
        - stored_at
        properties:
          stored_at:
            type: string
            format: date-time
...
    StoredDriftReport:
      allOf:
      - $ref: '#/components/schemas/DriftReport'
      - type: object
        required:
        - stored_at
        properties:
          stored_at:
            type: string
            format: date-time
...